requirements.txt
python-docx==0.8.11
Pillow==9.0.1 
pywin32==304
git@github.com:amfor/excel2img.git